Add menu builder helper
This helper allows the building of PureCSS-compatible menus complete
with their menu items.

This commit additionally adds the test app's database.yml and secrets.yml files
to allow the running of specs without having to add the files manually.
As it's only an app meant for running tests on, I don't see this as a problem.