Add a partial refresh capability
This commit adds the notion of a partial refresh.

A partial refresh is one that is tied to an ajax request, and partially replaces the content on the page.

For example, let’s say you have pagination set up with `remote: true`, meaning the request will be made via ajax. If you add the `js-partial-refresh` class to the pagination element, the new behaviour will be to show a loading indicator, replace the parent’s content with the returned data, then remove the loading indicator.

The parent wrapper can be defined by a `data-parent` attribute on the ‘js-partial-refresh` element, or it will be the closest `.portlet` or `#main-content`.